hotfix

+++ panTompkins.h
@@ -33,7 +33,7 @@
 #ifndef PAN_TOMPKINS
 #define PAN_TOMPKINS
 
-typedef unsigned long dataType;
+typedef int dataType;
 
 void panTompkins(char file_in[],char file_out[]);
 void init(char file_in[], char file_out[]);
